Implements flightData loading and get Point function to DroneFlightData

+++ src/flight_data/geo_types.h
@@ -5,8 +5,8 @@ namespace dfv {
      * @brief A coordinate in latitude, longitude and altitude.
      */
     struct Coordinate {
-        float lat;
-        float lon;
-        float alt; //!< Altitude in meters
+        double lat;
+        double lon;
+        double alt; //!< Altitude in meters
     };
 } // namespace dfv
